# Dict add StringComparer.Ordinal{IgnoreCase} non-randomized optimization