--- a/src/System.Private.CoreLib/shared/System/Collections/Generic/NonRandomizedStringEqualityComparer.cs
+++ b/src/System.Private.CoreLib/shared/System/Collections/Generic/NonRandomizedStringEqualityComparer.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Globalization;
 using System.Runtime.Serialization;
 
 namespace System.Collections.Generic
@@ -29,6 +30,67 @@ namespace System.Collections.Generic
         {
             // We are doing this to stay compatible with .NET Framework.
             info.SetType(typeof(GenericEqualityComparer<string>));
+        }
+    }
+
+    [Serializable]
+    internal sealed class NonRandomizedStringComparer : IEqualityComparer<string>, IComparer<string>, ISerializable
+    {
+        internal static IEqualityComparer<string> Default { get; } = new NonRandomizedStringComparer();
+
+        private NonRandomizedStringComparer() { }
+
+        public bool Equals(string x, string y) => string.Equals(x, y);
+
+        public int GetHashCode(string obj) => obj?.GetNonRandomizedHashCode() ?? 0;
+
+        public int Compare(string x, string y)
+        {
+            if (ReferenceEquals(x, y))
+                return 0;
+            if (x == null)
+                return -1;
+            if (y == null)
+                return 1;
+
+            return string.CompareOrdinal(x, y);
+        }
+
+        public void GetObjectData(SerializationInfo info, StreamingContext context)
+        {
+            info.SetType(typeof(OrdinalComparer));
+            info.AddValue("_ignoreCase", false); // Do not rename (binary serialization)
+        }
+    }
+
+    [Serializable]
+    internal sealed class NonRandomizedIgnoreCaseStringEqualityComparer : IEqualityComparer<string>, ISerializable
+    {
+        internal static IEqualityComparer<string> Default { get; } = new NonRandomizedIgnoreCaseStringEqualityComparer();
+
+        private NonRandomizedIgnoreCaseStringEqualityComparer() { }
+
+        public bool Equals(string x, string y)
+        {
+            if (ReferenceEquals(x, y))
+            {
+                return true;
+            }
+
+            if (x is null || y is null || x.Length != y.Length)
+            {
+                return false;
+            }
+
+            return CompareInfo.EqualsOrdinalIgnoreCase(ref x.GetRawStringData(), ref y.GetRawStringData(), x.Length);
+        }
+
+        public int GetHashCode(string obj) => obj?.GetNonRandomizedIgnoreCaseHashCode() ?? 0;
+
+        public void GetObjectData(SerializationInfo info, StreamingContext context)
+        {
+            info.SetType(typeof(OrdinalComparer));
+            info.AddValue("_ignoreCase", true); // Do not rename (binary serialization)
         }
     }
 }