+473% / x5.7 speed up for hashing 10 char `StringComparer.OrdinalIgnoreCase` ascii keys in used  `Dictionary<string, TValue>`

+116% / x2.2 speed up for hashing 10 char `StringComparer.Ordinal` keys in used  `Dictionary<string, TValue>`

Fast path for ascii `StringComparer.OrdinalIgnoreCase` before collisions become high, when it switches to randomised hashing.

Recognise `StringComparer.Ordinal` in addition to `EqualityComparer<string>.Default` for non-randomized optimization. 

[apisof.net](https://apisof.net/catalog/System.StringComparer.Ordinal) says `StringComparer.Ordinal` has 4.8% usage on NuGet.

`StringComparer.OrdinalIgnoreCase` version 7.8% usage on NuGet

Resolves dotnet/corefx#28834

For ascii input https://github.com/dotnet/coreclr/pull/19436#issuecomment-412384276
```
                      Method |                 Input |       Mean |          Op/s | Scaled |
---------------------------- |---------------------- |-----------:|--------------:|-------:|
     NonRandomizedIgnoreCase |                       |   2.543 ns | 393,239,898.3 |   0.16 |
 RandomizedOrdinalIgnoreCase |                       |  16.365 ns |  61,107,451.3 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |                     1 |   3.677 ns | 271,977,635.4 |   0.09 |
 RandomizedOrdinalIgnoreCase |                     1 |  39.840 ns |  25,100,712.8 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |                    12 |   3.681 ns | 271,684,906.0 |   0.09 |
 RandomizedOrdinalIgnoreCase |                    12 |  42.557 ns |  23,497,789.0 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |                   123 |   5.090 ns | 196,459,260.6 |   0.12 |
 RandomizedOrdinalIgnoreCase |                   123 |  42.889 ns |  23,315,996.0 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |                  1234 |   5.090 ns | 196,451,630.3 |   0.11 |
 RandomizedOrdinalIgnoreCase |                  1234 |  44.283 ns |  22,582,258.6 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |                 12345 |   6.220 ns | 160,783,055.6 |   0.14 |
 RandomizedOrdinalIgnoreCase |                 12345 |  45.369 ns |  22,041,264.4 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |                123456 |   6.226 ns | 160,609,477.0 |   0.13 |
 RandomizedOrdinalIgnoreCase |                123456 |  46.565 ns |  21,475,384.8 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |               1234567 |   7.879 ns | 126,919,960.7 |   0.17 |
 RandomizedOrdinalIgnoreCase |               1234567 |  47.166 ns |  21,201,895.8 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |              12345678 |   7.879 ns | 126,916,990.6 |   0.16 |
 RandomizedOrdinalIgnoreCase |              12345678 |  48.925 ns |  20,439,378.7 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |             123456789 |   8.995 ns | 111,169,928.0 |   0.18 |
 RandomizedOrdinalIgnoreCase |             123456789 |  49.784 ns |  20,086,888.1 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |            1234567890 |   8.995 ns | 111,173,636.8 |   0.17 |
 RandomizedOrdinalIgnoreCase |            1234567890 |  51.606 ns |  19,377,588.1 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |  12345678901234567890 |  16.321 ns |  61,271,112.6 |   0.25 |
 RandomizedOrdinalIgnoreCase |  12345678901234567890 |  66.322 ns |  15,077,973.9 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |  12345(...)67890 [40] |  30.431 ns |  32,861,074.5 |   0.29 |
 RandomizedOrdinalIgnoreCase |  12345(...)67890 [40] | 105.175 ns |   9,507,940.1 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase |  12345(...)67890 [80] |  59.670 ns |  16,758,837.1 |   0.37 |
 RandomizedOrdinalIgnoreCase |  12345(...)67890 [80] | 162.946 ns |   6,137,005.6 |   1.00 |
                             |                       |            |               |        |
     NonRandomizedIgnoreCase | 12345(...)67890 [160] | 118.589 ns |   8,432,479.1 |   0.41 |
 RandomizedOrdinalIgnoreCase | 12345(...)67890 [160] | 288.826 ns |   3,462,290.7 |   1.00 |
```

/cc @jkotas @danmosemsft @GrabYourPitchforks @stephentoub @safern