@jkotas 2018-08-11
Isn't it going to change behavior of  `Dictionary.Comparer` property?