@benaadams 2018-08-11
> Isn't it going to change behavior of `Dictionary.Comparer` property?

Hmm... yes. 

Should I get the `EqualityComparer<TKey>.Default` constructor to return `StringComparer.Ordinal` for string so they are the same instance?

That shouldn't interfere with intrinsic use of `EqualityComparer<TKey>.Default`  (or not work because of it)?