@jkotas 2018-08-11
I do not think that works. EqualityComparer.Default returns EqualityComparer. StringComparer does not inherit from EqualityComparer.