@benaadams 2018-08-11
New type to differentiate them? `NonRandomizedStringComparer`