@benaadams 2018-08-11
Added an `StringComparer.OrdinalIgnoreCase` optimization also /cc @ViktorHofer