@benaadams 2018-08-11
Collisions check
```
466544 words

466544 uppercased // words.Select(s => s.ToUpperInvariant()).Distinct().ToArray()
466544 lowercased // words.Select(s => s.ToLowerInvariant()).Distinct().ToArray(); 

933085 uppercase+lowercase // upperwords.Concat(lowerwords).Distinct().ToArray();
3 unchanged by casing

                            Method |        Type | No Collisions |  1 Coll |  2 Coll |  3 Coll |
---------------------------------- |------------ |-------------- |-------- |-------- |-------- |
GetNonRandomizedIgnoreCaseHashCode |       lower |        466136 |     204 |       - |       - |
GetNonRandomizedIgnoreCaseHashCode |       upper |        466136 |     204 |       - |       - |
GetNonRandomizedIgnoreCaseHashCode | upper+lower |             3 |  466133 |       - |     204 |
                                   |             |               |         |         |         |
    GetRemainingIgnoreCaseHashCode |       lower |        466164 |     190 |       - |       - |
    GetRemainingIgnoreCaseHashCode |       upper |        466164 |     190 |       - |       - |
    GetRemainingIgnoreCaseHashCode | upper+lower |             3 |  466161 |       - |     190 |
```