@benaadams 2018-08-11
/cc @vancem re: related optimization for `HashSet<T>`?